tugas3 dan update

+++ Praktikum/src/tugas2/Matematika.java
@@ -21,8 +21,8 @@ public class Matematika {
         return a * b;
     }
 
-    public int pembagian(int a, int b) {
-        return a / b;
+    public float pembagian(int a, int b) {
+        return (float) a / b;
     }
 }
