+++ Praktikum/src/tugas2/MatematikaBeraksi.java
@@ -15,12 +15,11 @@ public class MatematikaBeraksi {
         int hasilPertambahan = kalkulator.pertambahan(20, 10);
         int hasilPengurangan = kalkulator.pengurangan(10, 5);
         int hasilPerkalian = kalkulator.perkalian(10, 3);
-        int hasilPembagian = kalkulator.pembagian(21, 2);
+        float hasilPembagian = kalkulator.pembagian(21, 2);
 
         System.out.println("Pertambahan : 20 + 10 = " + hasilPertambahan);
         System.out.println("Pengurangan : 10 - 5 = " + hasilPengurangan);
         System.out.println("Perkalian : 10 * 3 = " + hasilPerkalian);
-        //Hasil 10 merupakan permintaan dari module
         System.out.println("Pembagian : 21 / 2 = " + hasilPembagian);
     }
 }
